requirements.txt
django
django-crispy-forms
django-gravatar2
django-markdown2
djangorestframework
drf_yasg
